Сделал поля компонентов приватными и разделил модуль utils
