+++ src/presenter/board-presenter.js
@@ -5,10 +5,14 @@ import PointView from '../view/point-view.js';
 import EditPointView from '../view/edit-point-view.js';
 
 export default class BoardPresenter {
-  #container;
-  #destinationsModel;
-  #offersModel;
-  #pointsModel;
+  #container = null;
+  #destinationsModel = null;
+  #offersModel = null;
+  #pointsModel = null;
+
+  #sortView = null;
+  #pointListView = null;
+  #points = [];
 
   constructor({ container, destinationsModel, offersModel, pointsModel }) {
     this.#container = container;
@@ -18,24 +22,24 @@ export default class BoardPresenter {
   }
 
   init() {
-    render(new SortView(), this.#container);
+    this.#sortView = new SortView();
+    this.#pointListView = new PointListView();
+    this.#points = [...this.#pointsModel.get()];
+    this.#renderBoard();
+  }
 
-    const eventList = new PointListView();
-    render(eventList, this.#container);
-
-    const points = [...this.#pointsModel.get()];
-    render(new EditPointView({
-      point: points[0],
-      pointDestination: this.#destinationsModel.getById(points[0].destination),
-      pointOffers: this.#offersModel.getByType(points[0].type)
-    }), eventList.element);
-
-    for (let i = 1; i < points.length; i++) {
-      const point = points[i];
-      const pointDestination = this.#destinationsModel.getById(point.destination);
-      const pointOffers = this.#offersModel.getByType(point.type);
-      const pointView = new PointView({ point, pointDestination, pointOffers });
-      render(pointView, eventList.element);
+  #renderBoard() {
+    render(this.#sortView, this.#container);
+    render(this.#pointListView, this.#container);
+    for (let i = 0; i < this.#points.length; i++) {
+      this.#renderPoint(this.#points[i]);
     }
+  }
+
+  #renderPoint(point) {
+    const pointDestination = this.#destinationsModel.getById(point.destination);
+    const pointOffers = this.#offersModel.getByType(point.type);
+    const pointView = new PointView({ point, pointDestination, pointOffers });
+    render(pointView, this.#pointListView.element);
   }
 }
