+++ src/presenter/filter-presenter.js
@@ -2,7 +2,7 @@ import { render } from '../framework/render.js';
 import FilterView from '../view/filter-view.js';
 
 export default class FilterPresenter {
-  #container;
+  #container = null;
 
   constructor({ container }) {
     this.#container = container;
