+++ src/utils/point.js
@@ -1,15 +1,5 @@
 import dayjs from 'dayjs';
 
-function getRandomInteger(a = 0, b = 1) {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-}
-
-function getRandomValue(array) {
-  return array[0, getRandomInteger(0, array.length - 1)];
-}
-
 function formatDateToShortDate(date){
   return dayjs(date).format('MMM DD');
 }
@@ -32,4 +22,4 @@ function formatDuration(date1, date2){
   return `${duration.minute()}M`;
 }
 
-export { getRandomInteger, getRandomValue, formatDateToShortDate, formatDateToDateTimeHTML, formatDateToDateTime, formatDateToTime, formatDuration };
+export { formatDateToShortDate, formatDateToDateTimeHTML, formatDateToDateTime, formatDateToTime, formatDuration };
